+++ examples/Cartesian/Cartesian.ino
@@ -1,10 +1,10 @@
-#include <Tle493d.h>
-
 /**
 * For basic setup just create a Tle493d() object. If you want to use the wake up mode please use Tle493d_w2b6(). Also
 * the setUpdateRate() method is slightly different for different variants
 */
 
+
+#include <Tle493d.h>
 
 Tle493d Tle493dMagnetic3DSensor = Tle493d();
